AutoMapping: `regions` layer now always defines input and output regions
Previously, you could choose to either define a single `regions` layer
or separate `regions_input` and `regions_output` layers. Often though,
the regions will be the same for many rules and exceptions need to be
made only for some rules.

Now, all three regions layers can be used at the same time, and any tile
that is placed in the `regions` layer will extend both the input and the
output region. This avoids the need to keep the shared parts of the
`regions_input` and `regions_output` synchronized, while still enabling
per-rule exceptions.

Thanks to @stanleylkm for this suggestion!